Checkpoint native numeric carrier and provider corrections

+++ test/cli-build/runtime/nodejs/process-buffer-crypto-os.test.mjs
@@ -83,7 +83,7 @@ test("CLI emits expanded process operations from selected Node provider-package 
   assert.match(generatedSource, /Tsonic\.CSharp\.Node\.process\.exit\(Tsonic\.CSharp\.Generated\.__TsonicConversions\.LiftNullable<double, int>\(code, System\.Convert\.ToInt32\)\);/);
   assert.match(generatedSource, /return Tsonic\.CSharp\.Node\.process\.kill\(Tsonic\.CSharp\.Node\.process\.pid, 0\);/);
   assert.match(generatedSource, /return Tsonic\.CSharp\.Node\.process\.availableMemory\(\) \+ Tsonic\.CSharp\.Node\.process\.constrainedMemory\(\);/);
-  assert.match(generatedSource, /double\[\] parts = Tsonic\.CSharp\.Node\.process\.hrtime\(\);/);
+  assert.match(generatedSource, /long\[\] parts = Tsonic\.CSharp\.Node\.process\.hrtime\(\);/);
   assert.match(generatedSource, /return parts\[0\] \+ parts\[1\];/);
   assert.doesNotMatch(generatedSource, /return process\./);
   assert.doesNotMatch(generatedSource, /__unsupported/);
